prepublish must be told which version now

+++ prepublish.sh
@@ -1,6 +1,11 @@
-npm install --ignore-scripts canvas@$npm_package_version && {
-  rm -rf canvas;
-  mv node_modules/canvas canvas;
-  rm -r node_modules;
+if [ "$1" = "" ]; then
+  echo "Specify which version"
+  exit 1
+fi
+
+npm install --ignore-scripts canvas@$1 && {
+  rm -rf canvas
+  mv node_modules/canvas canvas
+  rm -r node_modules
 }
